Fix gcc warnings

+++ src/MathLib/Gauss.cpp
@@ -133,11 +133,11 @@ matrix_t* kernel_apply_to_matrix(matrix_t* dst, matrix_t* src, kernel_t* k) {
     size_t k2 = k->size / 2;
     
 #pragma omp parallel for
-    for (int j = 0; j < dst->cols; ++j) {
-        for (int i = 0; i < dst->rows; ++i) {
+    for (int j = 0; j < static_cast<int>(dst->cols); ++j) {
+        for (int i = 0; i < static_cast<int>(dst->rows); ++i) {
             double d = 0.0;
 
-            for (int n = 0; n < k->size; n++) {
+            for (int n = 0; n < static_cast<int>(k->size); n++) {
                 int p = i + n - k2;
                 p = normalize_index(p, dst->cols, k->mode);
                 d += src->data[p + j * src->cols] * k->data[n];
@@ -148,11 +148,11 @@ matrix_t* kernel_apply_to_matrix(matrix_t* dst, matrix_t* src, kernel_t* k) {
     }
     
 #pragma omp parallel for
-    for (int j = 0; j < dst->cols; ++j) {
-        for (int i = 0; i < dst->rows; ++i) {
+    for (int j = 0; j < static_cast<int>(dst->cols); ++j) {
+        for (int i = 0; i < static_cast<int>(dst->rows); ++i) {
             double d = 0.0;
 
-            for (int n = 0; n < k->size; ++n) {
+            for (int n = 0; n < static_cast<int>(k->size); ++n) {
                 int p = i + n - k2;
                 p = normalize_index(p, dst->rows, k->mode);
                 d += tmp->data[j + p * tmp->cols] * k->data[n];
@@ -198,11 +198,11 @@ texture_t* kernel_apply_to_texture(texture_t* t, kernel_t* k) {
     size_t k2 = k->size / 2;
     
 #pragma omp parallel for
-    for (int j = 0; j < t->size; ++j) {
-        for (int i = 0; i < t->size; ++i) {
+    for (int j = 0; j < static_cast<int>(t->size); ++j) {
+        for (int i = 0; i < static_cast<int>(t->size); ++i) {
             double r = 0.0, g = 0.0, b = 0.0;
 
-            for (int n = 0; n < k->size; n++) {
+            for (int n = 0; n < static_cast<int>(k->size); n++) {
                 int p = i + n - k2;
                 p = normalize_index(p, t->size, k->mode);
 
@@ -218,11 +218,11 @@ texture_t* kernel_apply_to_texture(texture_t* t, kernel_t* k) {
     }
     
 #pragma omp parallel for
-    for (int j = 0; j < t->size; ++j) {
-        for (int i = 0; i < t->size; ++i) {
+    for (int j = 0; j < static_cast<int>(t->size); ++j) {
+        for (int i = 0; i < static_cast<int>(t->size); ++i) {
             double r = 0.0, g = 0.0, b = 0.0;
 
-            for (int n = 0; n < k->size; ++n) {
+            for (int n = 0; n < static_cast<int>(k->size); ++n) {
                 int p = i + n - k2;
                 p = normalize_index(p, tmp->size, k->mode);
